remove "save" button that is used for saving input controls as a new report option (for CE servers).

+++ JasperMobile/Sources/ViewControllers/Repository&Libraries/ResourceViewer/JMResourceViewerViewController/ReportViewer/Filter Options/JMInputControlsViewController.m
@@ -568,10 +568,12 @@
 
 - (void)updateRightBurButtonItem
 {
-    if ([self currentReportOptionIsExisted]) {
-        self.navigationItem.rightBarButtonItem = [[UIBarButtonItem alloc] initWithImage:[UIImage imageNamed:@"delete_item"] style:UIBarButtonItemStyleDone target:self action:@selector(deleteReportOptionTapped:)];
-    } else {
-        self.navigationItem.rightBarButtonItem = [[UIBarButtonItem alloc] initWithImage:[UIImage imageNamed:@"save_item"] style:UIBarButtonItemStyleDone target:self action:@selector(createReportOptionTapped:)];
+    if ([JMUtils isServerProEdition]) {
+        if ([self currentReportOptionIsExisted]) {
+            self.navigationItem.rightBarButtonItem = [[UIBarButtonItem alloc] initWithImage:[UIImage imageNamed:@"delete_item"] style:UIBarButtonItemStyleDone target:self action:@selector(deleteReportOptionTapped:)];
+        } else {
+            self.navigationItem.rightBarButtonItem = [[UIBarButtonItem alloc] initWithImage:[UIImage imageNamed:@"save_item"] style:UIBarButtonItemStyleDone target:self action:@selector(createReportOptionTapped:)];
+        }
     }
 }
